feature DELETE by id(pk) created.

+++ src/routes/library.js
@@ -84,9 +84,11 @@ libraryRouter.post('/', async (req, res) => {
 // PUT
 // PUT by id(pk) // modify library
 libraryRouter.put('/:id', async (req, res) => {
-    let id = req.params.id;
+    let { id } = req.params;
     const { name, location, landline } = req.body;
-    if(!name || !location || !landline){
+    const params = { name, location, landline };
+
+    if(params.length <= 0){
         res
         .status(400)
         .json({err: 'missing parameters'})
@@ -94,9 +96,9 @@ libraryRouter.put('/:id', async (req, res) => {
 
         return;
     }
+
     Library.findByPk(id)
     .then(async library => {
-
         if(!library){
             res
             .status(404)
@@ -124,22 +126,69 @@ libraryRouter.put('/:id', async (req, res) => {
 // DELETE
 // DELETE ALL
 libraryRouter.delete('/', async (req, res) => {
-    try{
-        // delte all table content:
-        await Library.destroy({
-            truncate: true,
-        });
+    // we set delete all to false by default.
+    let { deleteAll = false, id } = req.body;
+    // we verify if we have the param deleteAll, and if we do, we set the scope variable to be true.
+    if(req.body.deleteAll === true){
+        deleteAll = true;
+    }
 
-        res
-        .status(200)
-        .json({message: 'all librarys have been deleted'})
-        .end()
-    }catch{
-        console.log('there was an error')
-        res
-        .status(400)
-        .json({err: 'internal server error'})
-        .end()
+    if(deleteAll === false){
+        // we verify if we have the param id
+        if(id !== undefined && typeof id === 'number'){
+            // we delete a specific one
+            try{
+                Library.destroy({
+                    where: {
+                        id: id,
+                    }
+                }).then(destroyedRows => {
+                    if(destroyedRows === 0){
+                        console.log('no library has been deleted')
+                        res
+                        .status(404)
+                        .json({err: `library with id ${id} not found`})
+                        .end();
+                    } else{
+                        console.log(`the library with id ${id} has been deleted`)
+                        res
+                        .status(200)
+                        .json({message: `the library with id ${id} has been deleted`})
+                        .end();
+                    }
+                })
+            } catch (err){
+                console.log('there was an error')
+                res
+                .status(400)
+                .json({err: 'internal server error'})
+                .end()
+            }
+        } else{
+            res
+            .status(400)
+            .json({err: 'missing parameters'})
+            .end()
+        }
+    } else if(deleteAll === true){
+
+        try {
+            // delte all table content:
+            await Library.destroy({
+                truncate: true,
+            });
+
+            res
+            .status(200)
+            .json({message: 'all libraries have been deleted'})
+            .end()
+        } catch {
+            res
+            .status(400)
+            .json({err: 'internal server error'})
+            .end()
+        }
+
     }
 })
 module.exports = { libraryRouter };
